+++ CHANGELOG.md
@@ -16,3 +16,4 @@
  * added function `faria.string.capitalise()`
  * added function `faria.data.cloneObject()`
  * added function `faria.data.shallowSort()`
+ * added function `faria.data.isInObjList()`